Split stats dir into multiple files
